+++ cookbooks/passenger-nginx/metadata.rb
@@ -1,7 +1,7 @@
-name             'nginx'
+name             'rails'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures nginx'
+description      'Installs/Configures rails'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.6'
+version          '0.1.0'